experimentation/requirements.txt
# Deep Research Agent - Experimentation Requirements
# Python 3.11+

# LLM Framework
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Tracing & Evaluation
langsmith>=0.1.0

# Web Search
tavily-python>=0.5.0
tavily

# HTTP Client (for FACT-Lite citation validation)
httpx>=0.27.0

# Environment & Config
python-dotenv>=1.0.0

# Data Processing
pyyaml>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

IPython